Use test pattern files for tests
