+++ tests/ImageSharp.Tests/Processing/Processors/Filters/BlackWhiteTest.cs
@@ -7,33 +7,33 @@ using SixLabors.ImageSharp.Tests.TestUtilities.ImageComparison;
 using SixLabors.Primitives;
 using Xunit;
 
-namespace SixLabors.ImageSharp.Tests.Processing.Processors.Effects
+namespace SixLabors.ImageSharp.Tests.Processing.Processors.Filters
 {
-    public class InvertTest : FileTestBase
+    public class BlackWhiteTest : FileTestBase
     {
         [Theory]
-        [WithFileCollection(nameof(DefaultFiles), DefaultPixelType)]
-        public void ImageShouldApplyInvertFilter<TPixel>(TestImageProvider<TPixel> provider)
+        [WithTestPatternImages(100, 100, DefaultPixelType)]
+        public void ImageShouldApplyBlackWhiteFilter<TPixel>(TestImageProvider<TPixel> provider)
             where TPixel : struct, IPixel<TPixel>
         {
             using (Image<TPixel> image = provider.GetImage())
             {
-                image.Mutate(x => x.Invert());
+                image.Mutate(x => x.BlackWhite());
                 image.DebugSave(provider);
             }
         }
 
         [Theory]
-        [WithFileCollection(nameof(DefaultFiles), DefaultPixelType)]
-        public void ImageShouldApplyInvertFilterInBox<TPixel>(TestImageProvider<TPixel> provider)
+        [WithTestPatternImages(100, 100, DefaultPixelType)]
+        public void ImageShouldApplyBlackWhiteFilterInBox<TPixel>(TestImageProvider<TPixel> provider)
             where TPixel : struct, IPixel<TPixel>
         {
             using (Image<TPixel> source = provider.GetImage())
             using (Image<TPixel> image = source.Clone())
             {
-                var bounds = new Rectangle(10, 10, image.Width / 2, image.Height / 2);
+                var bounds = new Rectangle(image.Width / 4, image.Width / 4, image.Width / 2, image.Height / 2);
 
-                image.Mutate(x => x.Invert(bounds));
+                image.Mutate(x => x.BlackWhite(bounds));
                 image.DebugSave(provider);
 
                 ImageComparer.Tolerant().VerifySimilarityIgnoreRegion(source, image, bounds);
